refactor(front): split runtime store module
