+++ frontend/src/views/JobsView.vue
@@ -11,7 +11,7 @@ import { onMounted, watch, computed } from 'vue'
 import { useRouter, useRoute } from 'vue-router'
 import type { LocationQueryRaw } from 'vue-router'
 import { useRuntimeStore } from '@/stores/runtime'
-import type { JobSortCriterion, JobSortOrder } from '@/stores/runtime'
+import type { JobSortCriterion, JobSortOrder } from '@/stores/runtime/jobs'
 import { useClusterDataPoller } from '@/composables/DataPoller'
 import { compareClusterJobSortOrder } from '@/composables/GatewayAPI'
 import type { ClusterJob } from '@/composables/GatewayAPI'
